move wal iterator to wal/slatedb module
