+++ slatedb/src/wal/slatedb/mod.rs
@@ -3,6 +3,7 @@
 pub(crate) mod admin;
 pub(crate) mod gc;
 pub(crate) mod reader;
+pub(crate) mod wal_iterator;
 pub(crate) mod wal_sst_builder;
 pub(crate) mod wal_writer;
 pub(crate) mod writer_init;
